api models (rename), get user by (key values, not hardcode)
